Rewrite Chapter 6
In short, the changes are as follows:
* I identify three kinds of sampling methods to
investigate sociotehcnologies of identification based
on issues of their scale as they extends over many spaces,
 actors, and time.
* I focus on multi-temporal sampling and identify
two heuristics to detect important moment based on their interpretative
flexibility and becoming part of an infrastructure (gateways).
* I now use the two heuristics to analyse the moments
in the evolution of the ELISE data matching software.
Essentialy, the empirical findings make it possible
to see how actors become entangled in logics of security,
and how knowledge about identification travels.

Sent to supervisors on 20 June 2022. See finding-blind-spots-20220620.docx